+++ projects/demo/src/app/demo/pages/common-doc/base-control/base-control.component.ts
@@ -1,8 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-
-@Component({
-    selector: 'pa-demo-base-control',
-    templateUrl: './base-control.component.html',
-    changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class BaseControlComponent {}