+++ projects/demo/src/app/demo/pages/common-doc/base-textfield/base-textfield.component.ts
@@ -1,8 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-
-@Component({
-    selector: 'pa-demo-base-textfield',
-    templateUrl: './base-textfield.component.html',
-    changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class BaseTextfieldComponent {}