+++ projects/demo/src/app/demo/pages/common-doc/base-textfield/deprecated-base-textfield.component.ts
@@ -1,0 +1,8 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+    selector: 'pa-demo-deprecated-base-textfield',
+    templateUrl: './deprecated-base-textfield.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class DeprecatedBaseTextfieldComponent {}