+++ projects/pastanaga-angular/src/lib/controls/textfield/input/input-component-tests/input-component-standalone.spec.ts
@@ -37,6 +37,7 @@ import {
     testType,
     testWriteValueDoNotEmit,
 } from './common-behaviors.spec';
+import { TextInputType, UpdateOnStrategy } from '../../../form-field.model';
 
 @Component({
     template: ` <pa-input
@@ -84,9 +85,9 @@ export class TestComponent {
     showAllErrors = false;
     errorMessages?: any;
     errorMessage?: string;
-    updateOn = 'change';
+    updateOn: UpdateOnStrategy = 'change';
     value?: string | number;
-    type?: string;
+    type?: TextInputType;
     placeholder?: string;
     required = false;
     pattern?: string | RegExp;