+++ projects/pastanaga-angular/src/public-api.ts
@@ -16,6 +16,7 @@ export * from './lib/breakpoint-observer/breakpoint.observer';
 
 export * from './lib/controls/deprecated-base-control.directive';
 export * from './lib/controls/control.model';
+export * from './lib/controls/form-field.model';
 export * from './lib/controls/textfield/deprecated-base-text-field.directive';
 export * from './lib/controls/textfield/text-field.module';
 export * from './lib/controls/textfield/select/deprecated-select.component';